danixsik_bot2/requirements.txt
aiogram==2.25.1
mutagen==1.46.0
Pillow==10.0.0
python-dotenv==1.0.0